CTFs/CAT Reloaded CTF/legacy devloper/legacy_developer/requirements.txt
Flask-SQLAlchemy
pyjwt[crypto]
Werkzeug
Jinja2
APScheduler
itsdangerous
Flask
requests